Moved workspace setup
